requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.0
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==4.0
django-allauth==0.46.0
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
idna==3.3
importlib-metadata==4.8.2
Markdown==3.3.6
oauthlib==3.1.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
zipp==3.6.0
